requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
redis
python-dotenv
